+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 *   bugfix fixing broken interactive ssh tutorial from docs
 *   change to automatic tty merging into a single pty if `_tty_in=True` and `_tty_out=True`
 *   introducing `_unify_ttys`, default False, which allows explicit tty merging into single pty
+*   contrib command for `ssh` connections requiring passwords
 
 ## 1.12.14 - 6/6/17
 *   bugfix for poor sleep performance [#378](https://github.com/amoffat/sh/issues/378)